requirements.txt
streamlit==1.31.0
google-cloud-bigquery==3.14.1
google-cloud-aiplatform==1.38.1
google-generativeai==0.3.2
google-auth==2.25.2
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
db-dtypes==1.2.0
pyarrow==14.0.2